+++ src/main/java/com/ycloud/client/api/WhatsappMessagesApi.java
@@ -75,8 +75,138 @@ public class WhatsappMessagesApi {
     }
 
     /**
+     * Build call for retrieve
+     * @param id ID of the object. (required)
+     * @param _callback Callback for upload/download progress
+     * @return Call to execute
+     * @throws ApiException If fail to serialize the request body object
+     * @http.response.details
+     <table summary="Response Details" border="1">
+        <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
+        <tr><td> 200 </td><td> Successfully retrieved the object. </td><td>  -  </td></tr>
+     </table>
+     */
+    public okhttp3.Call retrieveCall(String id, final ApiCallback _callback) throws ApiException {
+        String basePath = null;
+        // Operation Servers
+        String[] localBasePaths = new String[] {  };
+
+        // Determine Base Path to Use
+        if (localCustomBaseUrl != null){
+            basePath = localCustomBaseUrl;
+        } else if ( localBasePaths.length > 0 ) {
+            basePath = localBasePaths[localHostIndex];
+        } else {
+            basePath = null;
+        }
+
+        Object localVarPostBody = null;
+
+        // create path and map variables
+        String localVarPath = "/whatsapp/messages/{id}"
+            .replaceAll("\\{" + "id" + "\\}", localVarApiClient.escapeString(id.toString()));
+
+        List<Pair> localVarQueryParams = new ArrayList<Pair>();
+        List<Pair> localVarCollectionQueryParams = new ArrayList<Pair>();
+        Map<String, String> localVarHeaderParams = new HashMap<String, String>();
+        Map<String, String> localVarCookieParams = new HashMap<String, String>();
+        Map<String, Object> localVarFormParams = new HashMap<String, Object>();
+
+        final String[] localVarAccepts = {
+            "application/json"
+        };
+        final String localVarAccept = localVarApiClient.selectHeaderAccept(localVarAccepts);
+        if (localVarAccept != null) {
+            localVarHeaderParams.put("Accept", localVarAccept);
+        }
+
+        final String[] localVarContentTypes = {
+            
+        };
+        final String localVarContentType = localVarApiClient.selectHeaderContentType(localVarContentTypes);
+        if (localVarContentType != null) {
+            localVarHeaderParams.put("Content-Type", localVarContentType);
+        }
+
+        String[] localVarAuthNames = new String[] { "api_key" };
+        return localVarApiClient.buildCall(basePath, localVarPath, "GET", localVarQueryParams, localVarCollectionQueryParams, localVarPostBody, localVarHeaderParams, localVarCookieParams, localVarFormParams, localVarAuthNames, _callback);
+    }
+
+    @SuppressWarnings("rawtypes")
+    private okhttp3.Call retrieveValidateBeforeCall(String id, final ApiCallback _callback) throws ApiException {
+        
+        // verify the required parameter 'id' is set
+        if (id == null) {
+            throw new ApiException("Missing the required parameter 'id' when calling retrieve(Async)");
+        }
+        
+
+        okhttp3.Call localVarCall = retrieveCall(id, _callback);
+        return localVarCall;
+
+    }
+
+    /**
+     * Retrieve a WhatsApp message
+     * <p>
+     * Retrieves a WhatsApp message you&#39;ve previously sent.
+     * @param id ID of the object. (required)
+     * @return WhatsappMessage
+     * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
+     * @http.response.details
+     <table summary="Response Details" border="1">
+        <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
+        <tr><td> 200 </td><td> Successfully retrieved the object. </td><td>  -  </td></tr>
+     </table>
+     */
+    public WhatsappMessage retrieve(String id) throws ApiException {
+        ApiResponse<WhatsappMessage> localVarResp = retrieveWithHttpInfo(id);
+        return localVarResp.getData();
+    }
+
+    /**
+     * Retrieve a WhatsApp message
+     * <p>
+     * Retrieves a WhatsApp message you&#39;ve previously sent.
+     * @param id ID of the object. (required)
+     * @return ApiResponse&lt;WhatsappMessage&gt;
+     * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
+     * @http.response.details
+     <table summary="Response Details" border="1">
+        <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
+        <tr><td> 200 </td><td> Successfully retrieved the object. </td><td>  -  </td></tr>
+     </table>
+     */
+    public ApiResponse<WhatsappMessage> retrieveWithHttpInfo(String id) throws ApiException {
+        okhttp3.Call localVarCall = retrieveValidateBeforeCall(id, null);
+        Type localVarReturnType = new TypeToken<WhatsappMessage>(){}.getType();
+        return localVarApiClient.execute(localVarCall, localVarReturnType);
+    }
+
+    /**
+     * Retrieve a WhatsApp message (asynchronously)
+     * <p>
+     * Retrieves a WhatsApp message you&#39;ve previously sent.
+     * @param id ID of the object. (required)
+     * @param _callback The callback to be executed when the API call finishes
+     * @return The request call
+     * @throws ApiException If fail to process the API call, e.g. serializing the request body object
+     * @http.response.details
+     <table summary="Response Details" border="1">
+        <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
+        <tr><td> 200 </td><td> Successfully retrieved the object. </td><td>  -  </td></tr>
+     </table>
+     */
+    public okhttp3.Call retrieveAsync(String id, final ApiCallback<WhatsappMessage> _callback) throws ApiException {
+
+        okhttp3.Call localVarCall = retrieveValidateBeforeCall(id, _callback);
+        Type localVarReturnType = new TypeToken<WhatsappMessage>(){}.getType();
+        localVarApiClient.executeAsync(localVarCall, localVarReturnType, _callback);
+        return localVarCall;
+    }
+    /**
      * Build call for send
-     * @param whatsappMessageSendRequest  (optional)
+     * @param whatsappMessageSendRequest  (required)
      * @param _callback Callback for upload/download progress
      * @return Call to execute
      * @throws ApiException If fail to serialize the request body object
@@ -134,6 +264,11 @@ public class WhatsappMessagesApi {
     @SuppressWarnings("rawtypes")
     private okhttp3.Call sendValidateBeforeCall(WhatsappMessageSendRequest whatsappMessageSendRequest, final ApiCallback _callback) throws ApiException {
         
+        // verify the required parameter 'whatsappMessageSendRequest' is set
+        if (whatsappMessageSendRequest == null) {
+            throw new ApiException("Missing the required parameter 'whatsappMessageSendRequest' when calling send(Async)");
+        }
+        
 
         okhttp3.Call localVarCall = sendCall(whatsappMessageSendRequest, _callback);
         return localVarCall;
@@ -144,7 +279,7 @@ public class WhatsappMessagesApi {
      * Send a WhatsApp message
      * <p>
      * Sends an outbound WhatsApp message.
-     * @param whatsappMessageSendRequest  (optional)
+     * @param whatsappMessageSendRequest  (required)
      * @return WhatsappMessage
      * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
      * @http.response.details
@@ -162,7 +297,7 @@ public class WhatsappMessagesApi {
      * Send a WhatsApp message
      * <p>
      * Sends an outbound WhatsApp message.
-     * @param whatsappMessageSendRequest  (optional)
+     * @param whatsappMessageSendRequest  (required)
      * @return ApiResponse&lt;WhatsappMessage&gt;
      * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
      * @http.response.details
@@ -181,7 +316,7 @@ public class WhatsappMessagesApi {
      * Send a WhatsApp message (asynchronously)
      * <p>
      * Sends an outbound WhatsApp message.
-     * @param whatsappMessageSendRequest  (optional)
+     * @param whatsappMessageSendRequest  (required)
      * @param _callback The callback to be executed when the API call finishes
      * @return The request call
      * @throws ApiException If fail to process the API call, e.g. serializing the request body object