+++ src/main/java/com/ycloud/client/model/EventType.java
@@ -56,6 +56,16 @@ public enum EventType {
   WHATSAPP_MESSAGE_UPDATED("whatsapp.message.updated"),
   
   /**
+   * Occurs when a WhatsApp business account phone number&#39;s name has been approved or rejected.
+   */
+  WHATSAPP_PHONE_NUMBER_NAME_UPDATED("whatsapp.phone_number.name_updated"),
+  
+  /**
+   * Occurs when a WhatsApp business account phone number&#39;s quality-related status is updated, and the status changes to &#x60;GREEN&#x60;, &#x60;YELLOW&#x60;, or &#x60;RED&#x60;.
+   */
+  WHATSAPP_PHONE_NUMBER_QUALITY_UPDATED("whatsapp.phone_number.quality_updated"),
+  
+  /**
    * Occurs when a WhatsApp template is reviewed, and the status changes to &#x60;APPROVED&#x60; or &#x60;REJECTED&#x60;.
    */
   WHATSAPP_TEMPLATE_REVIEWED("whatsapp.template.reviewed");