+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.2.0 (2022-10-27)
+
+- Add [WhatsApp](https://www.ycloud.com/whatsapp) templates API.
+- Add [WhatsApp](https://www.ycloud.com/whatsapp) phone numbers API.
+
 ## 1.1.0 (2022-10-14)
 
 - Add [Verify](https://www.ycloud.com/verify) API.